Add tests from python module to tests root
