+++ tests/data/tohoku_shoreline_1min.npy
