+++ tests/test_dtopotools.py
@@ -5,12 +5,13 @@ import sys
 import shutil
 import tempfile
 import inspect
+import time
 
 import numpy
 import nose
 
-import clawpack.geoclaw.topotools as topotools
 import clawpack.geoclaw.dtopotools as dtopotools
+import clawpack.geoclaw.tests as tests
 
 # Set local test directory to get local files
 testdir = os.path.dirname(__file__)
@@ -132,11 +133,7 @@ def test_read_sift_make_dtopo(save=False):
 
     times = [1.]
     dtopo = fault.create_dtopography(x,y,times)
-    # print "max dz = ",dtopo.dz_list[-1].smax()
-    
-    # temp_path = tempfile.mkdtemp()
-    # try:
-    # Load (and save) test data and make the comparison
+
     test_data_path = os.path.join(testdir, "data", "sift_test_data.tt3")
     if save:
         dtopo.write(test_data_path, dtopo_type=3)
@@ -148,15 +145,6 @@ def test_read_sift_make_dtopo(save=False):
         % (dtopo.DZ.shape, compare_data.DZ.shape)
 
     assert numpy.allclose(compare_data.DZ, dtopo.DZ)
-
-    # except AssertionError as e:
-    #     test_name = inspect.stack()[1][-2][0][:-3]
-    #     test_dump_path = os.path.join(os.getcwd(), test_name)
-    #     shutil.mkdir(test_dump_path)
-    #     shutil.copy(temp_path, test_dump_path)
-    #     raise e
-    # finally:
-    #     shutil.rmtree(temp_path)
 
 
 def test_SubdividedPlaneFault_make_dtopo(save=False):
@@ -226,12 +214,10 @@ def test_dtopo_io():
             dtopo = dtopotools.DTopography(path=path)
 
             assert test_dtopo.DZ.shape == dtopo.DZ.shape, \
-                   "Shape of DZ not equal for topo_type = %s." % k
+                   "Shape of DZ not equal for topo_type = %s." % dtopo.topo_type
 
             assert numpy.allclose(test_dtopo.DZ, dtopo.DZ), \
                 "DZ not equal for %s" % path
-
-
 
     except AssertionError as e:
         test_dump_path = os.path.join(os.getcwd(), "test_dtopo_io")
@@ -335,6 +321,144 @@ def test_vs_old_dtopo():
         raise e
     finally:
         shutil.rmtree(temp_path)
+
+
+def test_dynamic_tohoku(verbose=False, plot=False):
+    r"""Test dynamic faulting via a Tohoku example"""
+
+    shoreline_fname = os.path.join(testdir, 'data', 'tohoku_shoreline_1min.npy')
+    shoreline_xy = numpy.load(shoreline_fname)
+
+    subfault_fname = os.path.join(testdir, 'data', 'tohoku_ucsb.txt')
+    fault = dtopotools.UCSBFault()
+    fault.read(subfault_fname)
+    fault.rupture_type = 'dynamic'
+
+    if plot:
+        import matplotlib.pyplot as plt
+
+        fault.plot_subfaults(slip_color=True)  # plot final slip
+        plt.show()
+
+    # seafloor deformation:
+    quick_test = True
+
+    if quick_test:
+        xlower = 140.
+        xupper = 146.
+        ylower = 35.
+        yupper = 41.
+        xylim = [xlower,xupper,ylower,yupper]
+
+        # dtopo parameters for 4 min resolution:
+        mx = int((xupper - xlower)*15 + 1)
+        my = int((yupper - ylower)*15 + 1)
+    else:
+        xlower = 135.
+        xupper = 150.
+        ylower = 30.
+        yupper = 45.
+        xylim = [xlower,xupper,ylower,yupper]
+
+        # dtopo parameters for 1 min resolution:
+        mx = int((xupper - xlower)*60 + 1)
+        my = int((yupper - ylower)*60 + 1)
+
+
+    x = numpy.linspace(xlower,xupper,mx)
+    y = numpy.linspace(ylower,yupper,my)
+
+    tmax = 0.
+    for s in fault.subfaults:
+        tmax = max(tmax, s.rupture_time + s.rise_time + s.rise_time_ending)
+    if verbose:
+        print "rupture ends at time ",tmax
+
+    times = numpy.linspace(0,tmax,10)
+    dtopo = fault.create_dtopography(x,y,times,verbose=True)
+
+    dz_final = dtopo.DZ[-1]
+    dz_max = dz_final.max()
+
+    if plot:
+        # Incorporate this function in dtopotools to replace animate_dz_colors?
+        def plot_subfaults_dz(t, fig=None):
+            if fig is None:
+                fig = plt.figure(figsize=(12,5))
+            else:
+                fig.clf()
+            ax1 = fig.add_subplot(121)
+            ax2 = fig.add_subplot(122)
+            fault.plot_subfaults(axes=ax1, slip_color=True, slip_time=t, xylim=xylim)
+            dtopo.plot_dz_colors(axes=ax2, t=t, cmax_dz=dz_max)
+            ax1.plot(shoreline_xy[:,0],shoreline_xy[:,1],'g')
+            ax2.plot(shoreline_xy[:,0],shoreline_xy[:,1],'g')
+            plt.axis(xylim)
+            fig.show()
+            return fig
+
+        dtopo.plot_dz_colors(t=tmax)
+        plt.show()
+            
+        fig = plt.figure(figsize=(12,5))
+
+        for t in list(numpy.linspace(0,150,16)) + [170,200]:
+            plot_subfaults_dz(t,fig)
+            plt.draw()
+            plt.show()
+            time.sleep(1)
+
+    temp_path = tempfile.mkdtemp()
+    try:
+        fname = os.path.join(temp_path, 'tohoku_ucsb_dynamic.tt3')
+        dtopo.write(fname, 3)
+    except Exception as e:
+        test_name = inspect.stack()[1][-2][0][:-3]
+        test_dump_path = os.path.join(os.getcwd(), test_name)
+        shutil.mkdir(test_dump_path)
+        shutil.copy(temp_path, test_dump_path)
+        raise e
+    finally:
+        shutil.rmtree(temp_path)
+
+    if verbose:
+        print 'Created ',fname
+
+
+def test_subdivided_plane_fault(verbose=False, plot=False):
+    r"""Test SubdividedPlaneFault class"""
+
+    # get a unit source fault plane as starting point:
+    sift_slip = {'acsza1':1.}
+    fault = dtopotools.SiftFault(sift_slip)
+    fault_plane = fault.subfaults[0]
+    Mo = fault_plane.Mo()
+    if verbose:
+        print "original Mo = ",Mo
+
+    fault2 = dtopotools.SubdividedPlaneFault(fault_plane, nstrike=5, ndip=3)
+    if verbose:
+        print "new Mo = ",fault2.Mo()
+    if plot:
+        import matplotlib.pyplot as plt
+        fault2.plot_subfaults(slip_color=True)
+        plt.show()
+
+    slip_function = lambda xi,eta: xi*(1-xi)*eta
+    fault2 = dtopotools.SubdividedPlaneFault(fault_plane, nstrike=5, ndip=3, 
+            slip_function=slip_function, Mo=Mo)
+    if verbose:
+        print "new Mo = ",fault2.Mo()
+    if plot:
+        fault2.plot_subfaults(slip_color=True)
+        plt.show()
+
+    fault2.subdivide(nstrike=20, ndip = 10, slip_function=slip_function, Mo=Mo)
+    if verbose:
+        print "with finer resolution, Mo = ",fault2.Mo()
+    if plot:
+        fault2.plot_subfaults(slip_color=True)
+        plt.show()
 
 
 if __name__ == "__main__":
